wip: move internal auth things into `auth::raw`
